refactor delete list funcs in mongo
