+++ mongod/structs.go
@@ -81,111 +81,6 @@ func (c *appRepository) Get(ctx context.Context) ([]*model.App, error) {
 	return apps, nil
 }
 
-func (c *appRepository) ListChannels(ctx context.Context) ([]*model.Channel, error) {
-
-	findOptions := options.Find()
-	findOptions.SetLimit(100)
-
-	var channels []*model.Channel
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-	// Define a filter to fetch documents with the "channel_name" field
-	filter := bson.M{"channel_name": bson.M{"$exists": true}}
-
-	cur, err := collection.Find(ctx, filter, findOptions)
-	if err != nil {
-		logrus.Fatal(err)
-		return nil, err
-	}
-
-	// Finding multiple documents returns a cursor
-	// Iterating through the cursor allows us to decode documents one at a time
-	for cur.Next(context.TODO()) {
-		// create a value into which the single document can be decoded
-		var elem model.Channel
-		if err := cur.Decode(&elem); err != nil {
-			logrus.Fatal(err)
-			return nil, err
-		}
-
-		channels = append(channels, &elem)
-	}
-
-	cur.Close(ctx)
-
-	return channels, nil
-}
-
-func (c *appRepository) ListPlatforms(ctx context.Context) ([]*model.Platform, error) {
-
-	findOptions := options.Find()
-	findOptions.SetLimit(100)
-
-	var platforms []*model.Platform
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-	// Define a filter to fetch documents with the "platform_name" field
-	filter := bson.M{"platform_name": bson.M{"$exists": true}}
-
-	cur, err := collection.Find(ctx, filter, findOptions)
-	if err != nil {
-		logrus.Fatal(err)
-		return nil, err
-	}
-
-	// Finding multiple documents returns a cursor
-	// Iterating through the cursor allows us to decode documents one at a time
-	for cur.Next(context.TODO()) {
-		// create a value into which the single document can be decoded
-		var elem model.Platform
-		if err := cur.Decode(&elem); err != nil {
-			logrus.Fatal(err)
-			return nil, err
-		}
-
-		platforms = append(platforms, &elem)
-	}
-
-	cur.Close(ctx)
-
-	return platforms, nil
-}
-
-func (c *appRepository) ListArchs(ctx context.Context) ([]*model.Arch, error) {
-
-	findOptions := options.Find()
-	findOptions.SetLimit(100)
-
-	var archs []*model.Arch
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-	// Define a filter to fetch documents with the "arch_id" field
-	filter := bson.M{"arch_id": bson.M{"$exists": true}}
-
-	cur, err := collection.Find(ctx, filter, findOptions)
-	if err != nil {
-		logrus.Fatal(err)
-		return nil, err
-	}
-
-	// Finding multiple documents returns a cursor
-	// Iterating through the cursor allows us to decode documents one at a time
-	for cur.Next(context.TODO()) {
-		// create a value into which the single document can be decoded
-		var elem model.Arch
-		if err := cur.Decode(&elem); err != nil {
-			logrus.Fatal(err)
-			return nil, err
-		}
-
-		archs = append(archs, &elem)
-	}
-
-	cur.Close(ctx)
-
-	return archs, nil
-}
-
 func (c *appRepository) GetAppByName(appName string, ctx context.Context) ([]*model.App, error) {
 
 	findOptions := options.Find()
@@ -219,116 +114,6 @@ func (c *appRepository) GetAppByName(appName string, ctx context.Context) ([]*mo
 	cur.Close(ctx)
 
 	return apps, nil
-}
-
-func (c *appRepository) DeleteApp(id primitive.ObjectID, ctx context.Context) ([]string, int64, error) {
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-
-	filter := bson.D{primitive.E{Key: "_id", Value: id}}
-
-	// Retrieve the document before deletion
-	var app *model.App
-	err := collection.FindOne(ctx, filter).Decode(&app)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return nil, 0, fmt.Errorf("no app found with ID %s", id)
-		}
-		return nil, 0, fmt.Errorf("error retrieving app with ID %s: %s", id, err.Error())
-	}
-
-	deleteResult, err := collection.DeleteOne(ctx, filter)
-	if err != nil {
-		logrus.Fatal(err)
-
-		return nil, 0, err
-	}
-
-	var links []string
-	for _, artifact := range app.Artifacts {
-		link := string(artifact.Link)
-		links = append(links, link)
-	}
-
-	return links, deleteResult.DeletedCount, nil
-}
-
-func (c *appRepository) DeleteChannel(id primitive.ObjectID, ctx context.Context) (int64, error) {
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-
-	filter := bson.D{primitive.E{Key: "_id", Value: id}}
-
-	// Retrieve the document before deletion
-	var channel *model.Channel
-	err := collection.FindOne(ctx, filter).Decode(&channel)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return 0, fmt.Errorf("no channel found with ID %s", id)
-		}
-		return 0, fmt.Errorf("error retrieving channel with ID %s: %s", id, err.Error())
-	}
-
-	deleteResult, err := collection.DeleteOne(ctx, filter)
-	if err != nil {
-		logrus.Fatal(err)
-
-		return 0, err
-	}
-
-	return deleteResult.DeletedCount, nil
-}
-
-func (c *appRepository) DeletePlatform(id primitive.ObjectID, ctx context.Context) (int64, error) {
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-
-	filter := bson.D{primitive.E{Key: "_id", Value: id}}
-
-	// Retrieve the document before deletion
-	var platform *model.Platform
-	err := collection.FindOne(ctx, filter).Decode(&platform)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return 0, fmt.Errorf("no platform found with ID %s", id)
-		}
-		return 0, fmt.Errorf("error retrieving platform with ID %s: %s", id, err.Error())
-	}
-
-	deleteResult, err := collection.DeleteOne(ctx, filter)
-	if err != nil {
-		logrus.Fatal(err)
-
-		return 0, err
-	}
-
-	return deleteResult.DeletedCount, nil
-}
-
-func (c *appRepository) DeleteArch(id primitive.ObjectID, ctx context.Context) (int64, error) {
-
-	collection := c.client.Database(c.config.Database).Collection("apps")
-
-	filter := bson.D{primitive.E{Key: "_id", Value: id}}
-
-	// Retrieve the document before deletion
-	var arch *model.Arch
-	err := collection.FindOne(ctx, filter).Decode(&arch)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return 0, fmt.Errorf("no arch found with ID %s", id)
-		}
-		return 0, fmt.Errorf("error retrieving arch with ID %s: %s", id, err.Error())
-	}
-
-	deleteResult, err := collection.DeleteOne(ctx, filter)
-	if err != nil {
-		logrus.Fatal(err)
-
-		return 0, err
-	}
-
-	return deleteResult.DeletedCount, nil
 }
 
 func (c *appRepository) Upload(ctxQuery map[string]interface{}, appLink, extension string, ctx context.Context) (interface{}, error) {
